Replace legacy bootstrap
Promote the new bootstrap module to the top
