--- jormungandr/src/network/protocols/mod.rs
@@ -1,1 +1,0 @@
-pub mod bootstrap;
